Move `TreeMatcher` into its own private module.
